python-requirements.txt
openai==1.43.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5